Refactor map endpoints
