+++ server/api.ts
@@ -11,45 +11,21 @@ import Team, { PopulatedTeam } from "./models/team";
 import Tournament from "./models/tournament";
 import TourneyMap from "./models/tourney-map";
 import User, { IUser } from "./models/user";
-import { Request } from "./types";
+import { getOsuApi, checkPermissions } from "./util";
+
+import mapRouter from "./api/map";
 
 import { addAsync } from "@awaitjs/express";
 const router = addAsync(express.Router());
 
-// Nuking the type definitions for osuApi by setting it to `any` for now
-// The types seem to be incorrectly defined for some things (e.g. MultiplayerScore)
-const osuApi = new osu.Api(process.env.OSU_API_KEY, { parseNumeric: true });
-
 const logger = pino();
+const osuApi = getOsuApi();
 const CONTENT_DIR = fs.readdirSync(`${__dirname}/../client/src/content`);
 
-const round = (num: number) => Math.round(num * 100) / 100;
-const formatTime = (time: number) =>
-  Math.floor(time / 60) + ":" + (time % 60 < 10 ? "0" : "") + Math.floor(time % 60);
-const scaleTime = (time: number, mod: string) =>
-  mod === "DT" ? (time * 2) / 3 : mod === "HT" ? (time * 3) / 2 : time;
-const scaleBPM = (bpm: number, mod: string) =>
-  mod === "DT" ? (bpm * 3) / 2 : mod === "HT" ? (bpm * 2) / 3 : bpm;
-const scaleDiff = (diff: number, mod: string) => {
-  if (mod === "HR" || mod === "HDHR") {
-    return Math.min(10, round(diff * 1.4));
-  }
-  if (mod == "EZ") {
-    return Math.min(10, round(diff / 2));
-  }
-  return diff;
-};
-
-const checkPermissions = (user: IUser, tourney: string, roles: string[]) => {
-  return (
-    user &&
-    user.username &&
-    (user.admin ||
-      user.roles.some(
-        (r) => ["Host", "Developer", ...roles].includes(r.role) && r.tourney == tourney
-      ))
-  );
-};
+// Parts of the API are gradually being split out into separate files
+// These are the sub-routers that have been migrated/refactored
+router.use(mapRouter);
+// ----------------------
 
 const isAdmin = (user: IUser, tourney: string) => checkPermissions(user, tourney, []);
 const canViewHiddenPools = (user: IUser, tourney: string) =>
@@ -134,91 +110,6 @@ const parseWarmup = async (warmup: string) => {
 
   return `https://osu.ppy.sh/beatmapsets/${mapData.beatmapSetId}#taiko/${warmupMapId}`;
 };
-
-/**
- * POST /api/map
- * Registers a new map into a mappool
- * Params:
- *   - id: ID of the map
- *   - mod: mod of the map
- *   - index: e.g. 3 for NM3, HD3, HR3
- *   - tourney: identifier for the tourney
- *   - stage: which pool, e.g. qf, sf, f, gf
- * Returns the newly-created Map document
- */
-router.postAsync("/map", ensure.isPooler, async (req: Request, res: Response) => {
-  logger.info(`${req.user.username} added ${req.body.id} to ${req.body.stage} mappool`);
-
-  const mod = req.body.mod;
-  const modId = { EZ: 2, HR: 16, HDHR: 16, DT: 64, HT: 256 }[mod] || 0; // mod enum used by osu api
-  const mapData = (await osuApi.getBeatmaps({ b: req.body.id, mods: modId, m: 1, a: 1 }))[0];
-
-  // all map metadata cached in our db, so we don't need to spam calls to the osu api
-  const newMap = new TourneyMap({
-    ...req.body,
-    mapId: parseInt(mapData.id),
-    title: mapData.title,
-    artist: mapData.artist,
-    creator: mapData.creator,
-    diff: mapData.version,
-    bpm: round(scaleBPM(mapData.bpm, mod)),
-    sr: round(mapData.difficulty.rating),
-    od: scaleDiff(mapData.difficulty.overall, mod),
-    hp: scaleDiff(mapData.difficulty.drain, mod),
-    length: formatTime(scaleTime(mapData.length.total, mod)),
-    image: `https://assets.ppy.sh/beatmaps/${mapData.beatmapSetId}/covers/cover.jpg`,
-    pooler: req.user.username,
-  });
-  await newMap.save();
-  res.send(newMap);
-});
-
-/**
- * GET /api/maps
- * Get all the maps for a given mappool (if the user has access)
- * Params:
- *   - tourney: identifier for the tourney
- *   - stage: which pool, e.g. qf, sf, f, gf
- */
-router.getAsync("/maps", async (req, res) => {
-  const [tourney, maps] = await Promise.all([
-    Tournament.findOne({ code: req.query.tourney }),
-    TourneyMap.find({ tourney: req.query.tourney, stage: req.query.stage }),
-  ]);
-
-  // if super hacker kiddo tries to view a pool before it's released
-  const stageData = tourney.stages.filter((s) => s.name === req.query.stage)[0];
-  if (!stageData.poolVisible && !canViewHiddenPools(req.user, req.query.tourney)) {
-    return res.status(403).send({ error: "This pool hasn't been released yet!" });
-  }
-
-  const mods = { NM: 0, HD: 1, HR: 2, DT: 3, FM: 4, HT: 5, HDHR: 6, EZ: 7, CV: 8, EX: 9, TB: 10 };
-  maps.sort((a, b) => {
-    if (mods[a.mod] - mods[b.mod] != 0) {
-      return mods[a.mod] - mods[b.mod];
-    }
-    return a.index - b.index;
-  });
-  res.send(maps);
-});
-
-/**
- * DELETE /api/maps
- * Delete a map from the pool
- * Params:
- *   - id: _id of the map to delete
- *   - tourney: identifier for the tourney
- *   - stage: which pool, e.g. qf, sf, f, gf
- */
-router.deleteAsync("/map", ensure.isPooler, async (req, res) => {
-  logger.info(`${req.user.username} deleted ${req.body.id} from ${req.body.stage} pool`);
-  await TourneyMap.deleteOne({
-    tourney: req.body.tourney,
-    stage: req.body.stage,
-    _id: req.body.id,
-  });
-  res.send({});
-});
 
 /**
  * GET /api/whoami
